+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [api]: Implemented a `prune()` method for pruning a tree down to only the selected samples (0.4.0.dev1).
 * [api]: Fixed bug where `query.join_tree()` was pruning the original tree (instead of a copy) (0.4.0.dev1).
 * [cli]: Fixed bug where `--features-summary mlst` would not work in command-line interface (0.4.0.dev1).
+* [api]: Changed default NodeStyle for rendering trees such that nodes have size 0 (avoids inflating distances when many samples have distance 0). (0.4.0.dev2)
 
 # 0.3.0